created scanner file
